#328: Moved the items models to model/items
-
