+++ src/app/ffbe/model/items/item-category.model.ts
