+++ src/app/ffbe/model/items/item-recipe.model.ts
